requirement.txt
langchain
python-dotenv
langchain-google-genai
langchain_community 
faiss-cpu